Manejo de excepciones y handler de exceptiones

+++ src/main/java/com/sportlife/user/controller/UserController.java
@@ -18,6 +18,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
+import com.sportlife.exception.ModelNotFoundException;
 import com.sportlife.user.dto.UserDTO;
 import com.sportlife.user.model.User;
 import com.sportlife.user.service.IUserService;
@@ -25,48 +26,64 @@ import com.sportlife.user.service.IUserService;
 @RestController
 @RequestMapping("/users")
 public class UserController {
-	//cuando nos comuniquemos entre capa ir con la interface para se sea desacoplado el codigo test
+	// cuando nos comuniquemos entre capa ir con la interface para se sea
+	// desacoplado el codigo test
 	@Autowired
 	private IUserService iUserService;
-	
+
 	@Autowired
 	private ModelMapper modelMapper;
 
 	@GetMapping("/{id}")
-	public ResponseEntity<UserDTO> findById(@PathVariable("id") Integer id) {
-		return new ResponseEntity<>(modelMapper.map(iUserService.findById(id), UserDTO.class), HttpStatus.OK);
+	public ResponseEntity<UserDTO> findById(@PathVariable("id") Integer id) throws ModelNotFoundException {
+		UserDTO userDto;
+		User user = iUserService.findById(id);
+		if (user == null) {
+			throw new ModelNotFoundException("ID NOT FOUND EXCEPTION: " + id);
+		} else {
+			userDto = modelMapper.map(user, UserDTO.class);
+		}
+		return new ResponseEntity<>(userDto, HttpStatus.OK);
 	}
 
 	@GetMapping
 	public ResponseEntity<List<UserDTO>> findAll() {
-		List<UserDTO> userDTOs= this.mapList(iUserService.findAll(), UserDTO.class);
+		List<UserDTO> userDTOs = this.mapList(iUserService.findAll(), UserDTO.class);
 		return new ResponseEntity<>(userDTOs, HttpStatus.OK);
 	}
 
 	@PostMapping
 	public ResponseEntity<Void> save(@RequestBody UserDTO userDto) {
-		
-		User userSave= iUserService.save(modelMapper.map(userDto, User.class));
-		URI uriUserLocation = ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}").buildAndExpand(userSave.getIdUser()).toUri();
+
+		User userSave = iUserService.save(modelMapper.map(userDto, User.class));
+		URI uriUserLocation = ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}")
+				.buildAndExpand(userSave.getIdUser()).toUri();
 		return ResponseEntity.created(uriUserLocation).build();
-		//return new ResponseEntity<>(iUserService.save(user), HttpStatus.CREATED);
+		// return new ResponseEntity<>(iUserService.save(user), HttpStatus.CREATED);
 	}
 
 	@PutMapping
-	public ResponseEntity<User> update(@RequestBody UserDTO userDto) {
+	public ResponseEntity<User> update(@RequestBody UserDTO userDto) throws ModelNotFoundException {
+		User user = iUserService.findById(userDto.getIdUser());
+		if (user == null) {
+			throw new ModelNotFoundException("ID NOT FOUND EXCEPTION: " + userDto.getIdUser());
+		}
 		return new ResponseEntity<>(iUserService.update(modelMapper.map(userDto, User.class)), HttpStatus.OK);
 	}
 
 	@DeleteMapping("/{id}")
-	public ResponseEntity<Void> deleteById(@PathVariable("id") Integer id) {
-		iUserService.delete(id);
+	public ResponseEntity<Void> deleteById(@PathVariable("id") Integer id) throws ModelNotFoundException {
+		User user = iUserService.findById(id);
+		if (user == null) {
+			throw new ModelNotFoundException("ID NOT FOUND EXCEPTION: " + id);
+		} else {
+			iUserService.delete(id);
+		}
 		return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 	}
-	
-	public  <S, T> List<T> mapList(List<S> source, Class<T> targetClass) {
-	    return source.stream().map(element -> modelMapper.map(element, targetClass))
-	      .collect(Collectors.toList());
+
+	public <S, T> List<T> mapList(List<S> source, Class<T> targetClass) {
+		return source.stream().map(element -> modelMapper.map(element, targetClass)).collect(Collectors.toList());
 	}
 
-	
 }
